разбил на h и с, сделал makefile, добавил оптимизации сортировки вставками и их график
