+++ lab3/.gitignore
@@ -2,3 +2,4 @@
 *.in
 *.out
 *.vscode
+*.txt
