+++ lab3/Makefile
@@ -1,0 +1,13 @@
+input_sort = insertion_sort
+
+CC = gcc
+
+PROGRAMS = tester.c testing.c $(input_sort).c
+
+all: tester
+
+tester:  $(PROGRAMS)
+	$(CC) $(PROGRAMS) -o tester
+	
+clean:
+	rm -rf *.o tester
