+++ lab3/graphics/intro_sort_recursion_depth.png
