+++ lab3/graphics/merge_sort.png
