+++ lab3/graphics/pyramid_sort_k.png
