+++ lab3/graphics/quadratic_sorts.png
