+++ lab3/graphics/quick_sort_tmd.png
