+++ lab3/graphics/quick_sorts_big_tests.png
