+++ lab3/graphics/quick_sorts_pivots.png
