+++ lab3/msd_sort.c
@@ -1,0 +1,69 @@
+#include "msd_sort.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <assert.h>
+#include <string.h>
+ 
+
+void msd_sort_rec(int* array, size_t n, int* result, int shift) {
+
+    assert(array);
+    assert(result);
+
+    if (n <= 1 || shift < 0) {
+        
+        return;
+    } 
+    
+    int count[BYTE_SIZE + 1] = {};
+    int mask = 0xff;
+    
+    for (size_t i = 0; i < n; ++i) {
+
+        int byteValue = ((array[i] >> shift) & mask) + 1;
+        count[byteValue]++;
+    }
+    
+    for (int i = 1; i < BYTE_SIZE + 1; ++i) {
+
+        count[i] += count[i - 1];
+    }
+    
+    for (int i = n - 1; i >= 0; --i) {
+
+        int byteValue = ((array[i] >> shift) & mask) + 1;
+        result[--count[byteValue]] = array[i];
+    }
+    
+    for (size_t i = 0; i < n; ++i) {
+
+        array[i] = result[i];
+    }
+    
+    size_t start = 0;
+
+    for (int i = 1; i < BYTE_SIZE + 1; ++i) {
+
+        if (count[i] > start) {
+            
+            msd_sort_rec(array + start, count[i] - start, result, shift - 8);
+            start = count[i];
+        }
+    }
+}
+
+void msd_sort(int* array, int n) {
+
+    assert(array);
+
+    int* result = (int*) calloc(n, sizeof(int));
+    
+    if (!result) {
+        
+        return; 
+    } 
+    
+    msd_sort_rec(array, n, result, 8);
+    
+    free(result);
+}
