--- lab3/testing.cpp
@@ -1,13 +1,0 @@
-#include "testing.h"
-#include <stdio.h>
-#include <stdlib.h>
-#include <assert.h>
-
-
-void test_sortings(const char* test_folder, void (*sort) (int*, int), const char* result_path, int from, int to, int step) {
-
-    for (int i = from; i < to; i += step) {
-
-        FILE* file_test = fopen();
-    }
-}
